+++ lib/src/api/moneroo_api_wrapper.dart
@@ -5,6 +5,7 @@ import 'package:moneroo_flutter_sdk/src/commons/enums.dart';
 import 'package:moneroo_flutter_sdk/src/commons/exceptions.dart';
 import 'package:moneroo_flutter_sdk/src/models/api_response.dart';
 import 'package:moneroo_flutter_sdk/src/models/customer.dart';
+import 'package:moneroo_flutter_sdk/src/models/methods.dart';
 import 'package:moneroo_flutter_sdk/src/models/payment.dart';
 import 'package:moneroo_flutter_sdk/src/models/payment_infos.dart';
 
@@ -44,6 +45,9 @@ class MonerooApi {
     String? description,
     String? callbackUrl,
     Map<String, dynamic>? metadata,
+
+    /// For custom methods
+    List<MonerooMethod>? methods,
   }) async {
     try {
       final res = await _dio.post<Map<String, dynamic>>(
@@ -54,12 +58,16 @@ class MonerooApi {
           'description': description ?? 'From Moneroo Flutter package',
           'customer': customer.toJson(),
           'return_url': callbackUrl ?? 'https://example.com',
+          'metadata': metadata,
+          'methods': methods?.map<String>((method) => method.name).toList(),
         },
       );
 
       final apiResponse = MonerooApiResponse.fromJson(res.data!);
 
-      return MonerooPayment.fromJson(apiResponse.data);
+      return MonerooPayment.fromJson(
+        apiResponse.data as Map<String, dynamic>,
+      );
     } on DioException catch (e) {
       if (e.response == null) throw ServiceUnavailableException();
 
@@ -80,14 +88,15 @@ class MonerooApi {
     required String paymentId,
   }) async {
     try {
-      // ignore: inference_failure_on_function_invocation
       final res = await _dio.get<Map<String, dynamic>>(
         '/${apiVersion.name}${Endpoints.payments}/$paymentId${Endpoints.verify}',
       );
 
       final apiResponse = MonerooApiResponse.fromJson(res.data!);
 
-      return MonerooPaymentInfos.fromJson(apiResponse.data);
+      return MonerooPaymentInfos.fromJson(
+        apiResponse.data as Map<String, dynamic>,
+      );
     } on DioException catch (e) {
       if (e.response == null) throw ServiceUnavailableException();
 
@@ -102,23 +111,37 @@ class MonerooApi {
     }
   }
 
-  ///
-  // void handleException(DioException e) {
-  //   if (e.response?.statusCode == 422) {
-  //     throw InvalidPayloadException(
-  //       errors: (e.response?.data as Map<String, dynamic>)['errors'],
-  //       message:
-  //           (e.response?.data as Map<String, dynamic>)['message'] as String,
-  //     );
-  //   }
+  /// Retrive all the available methods on Moneroo
+  Future<List<MonerooRemoteMethod>> getMonerooPaymentMethods() async {
+    final methods = <MonerooRemoteMethod>[];
 
-  //   if (e.response?.statusCode == 401) {
-  //     throw UnauthorizedException(
-  //       errors: (e.response?.data as Map<String, dynamic>)['errors'],
-  //       message:
-  //           (e.response?.data as Map<String, dynamic>)['message'] as String,
-  //     );
-  //   }
+    try {
+      final res = await _dio.get<Map<String, dynamic>>(
+        Endpoints.methods,
+      );
 
-  // }
+      final apiResponse = MonerooApiResponse.fromJson(res.data!);
+
+      for (final data in apiResponse.data as List<dynamic>) {
+        methods.add(
+          MonerooRemoteMethod.fromJson(
+            data as Map<String, dynamic>,
+          ),
+        );
+      }
+
+      return methods;
+    } on DioException catch (e) {
+      if (e.response == null) throw ServiceUnavailableException();
+
+      final error = MonerooException(
+        code: e.response!.statusCode,
+        errors: (e.response?.data as Map<String, dynamic>)['errors'],
+        message:
+            (e.response?.data as Map<String, dynamic>)['message'] as String,
+      );
+
+      throw error;
+    }
+  }
 }