+++ test/src/moneroo_api_test.dart
@@ -5,20 +5,33 @@ import 'package:moneroo_flutter_sdk/moneroo_flutter_sdk.dart';
 
 void main() {
   final api = MonerooApi(
-    apiKey: 'YOUR_API_KEY',
+    apiKey: 'API-KEY',
   );
 
   group('MonerooApi', () {
     late String paymentId;
 
+    test('can get all the methods', () async {
+      final payment = await api.getMonerooPaymentMethods();
+
+      expect(
+        payment,
+        isNotEmpty,
+      );
+    });
+
     test('can init payment', () async {
       final payment = await api.initPayment(
-        amount: 1,
+        amount: 100,
         customer: MonerooCustomer(
           email: 'junior.medehou00@gmail.com',
           firstName: 'Elikem',
           lastName: 'Medehou',
         ),
+        metadata: {
+          'key': 'order_id',
+          'value': '1234567890',
+        },
       );
 
       paymentId = payment.id;